requirements.txt
streamlit
tensorflow==2.15.0
opencv-python-headless
pillow
pandas
numpy